requirements.txt
cachetools==3.1.0
certifi==2018.11.29
chardet==3.0.4
dialogflow==0.5.2
google-api-core==1.8.0
google-auth==1.6.3
googleapis-common-protos==1.5.8
grpcio==1.19.0
idna==2.8
praw==6.0.0
prawcore==1.0.1
protobuf==3.7.0
pyasn1==0.4.5
pyasn1-modules==0.2.4
pytz==2018.9
PyYAML==3.13
requests==2.21.0
rsa==4.0
six==1.12.0
update-checker==0.16
urllib3==1.24.1
